[kv_offload]: Add request_finished method to OffloadingManager and decouple store policy
Adds a `request_finished` method to `OffloadingManager` so implementations can
react when a request ends, e.g. to flush a deferred transfer for the last partial
block. The method is a no-op default to keep existing subclasses compatible.
`FilterReusedOffloadingManager` delegates it to its backing manager, and the
connector scheduler calls it alongside the new policy hook.
The block-selection logic embedded in `_build_store_jobs` is pulled out
into a new `OffloadPolicy` class. The existing behaviour becomes
`StoreOnComputePolicy`, which owns the per-request, per-group progress
index that previously lived as `next_stored_block_idx` on `RequestGroupState`.
The renamed `RequestKVState` (was `RequestOffloadState`) now only tracks
KV state — offload keys, block IDs, in-flight jobs.

The scheduler state types are moved to a new `state.py` so that
`policy.py` can import them without creating a circular dependency back
through `scheduler.py`.
